Waring fixes for time_t - fix for tests failure

+++ fpmsyncd/fpmsyncd.cpp
@@ -1,4 +1,5 @@
 #include <iostream>
+#include <inttypes.h>
 #include "logger.h"
 #include "select.h"
 #include "selectabletimer.h"
@@ -149,7 +150,7 @@ int main(int argc, char **argv)
                         if (eoiuFlagsSet(bgpStateTable))
                         {
                             /* Obtain eoiu hold timer defined for bgp docker */
-                            uint32_t eoiuHoldIval = WarmStart::getWarmStartTimer("eoiu_hold", "bgp");
+                            uintmax_t eoiuHoldIval = WarmStart::getWarmStartTimer("eoiu_hold", "bgp");
                             if (!eoiuHoldIval)
                             {
                                 eoiuHoldTimer.setInterval(timespec{DEFAULT_EOIU_HOLD_INTERVAL, 0});
@@ -161,7 +162,7 @@ int main(int argc, char **argv)
                             }
                             eoiuHoldTimer.start();
                             s.addSelectable(&eoiuHoldTimer);
-                            SWSS_LOG_NOTICE("Warm-Restart started EOIU hold timer which is to expire in %d seconds.", eoiuHoldIval);
+                            SWSS_LOG_NOTICE("Warm-Restart started EOIU hold timer which is to expire in %" PRIuMAX " seconds.", eoiuHoldIval);
                             s.removeSelectable(&eoiuCheckTimer);
                             continue;
                         }
